feat: remove dm driver
