+++ server/controller/db/metadb/config/session.go
@@ -1,0 +1,27 @@
+/*
+ * Copyright (c) 2024 Yunshan Networks
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package config
+
+type SessionConfig struct {
+	DBCfg Config
+	// set UseDabase=false in dsn only when migrating Metadb
+	UseDabase bool
+	// set TimeoutCoefficient=2 in dsn only when migrating Metadb
+	TimeoutCoefficient uint16
+	// set multiStatements=true in dsn only when migrating Metadb
+	MultiStatements bool
+}
