+++ server/controller/db/metadb/session/edition/go.mod
@@ -1,0 +1,3 @@
+module github.com/deepflowio/deepflow/server/controller/db/metadb/session/edition
+
+go 1.24.0
